requirements.txt
ntplib==0.3.4
tcp_latency==0.0.8
PyAutoGUI==0.9.53
python_dateutil==2.8.2
opencv-python>=4.5.4.60
